requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
werkzeug==2.0.1
jinja2==3.0.1
markupsafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
bootstrap-flask==2.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
pytest
pytest-cov
